+++ lib/screens/signal/signal_list.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:periodik/providers/points_provider.dart';
+import 'package:periodik/screens/point/delete_point.dart';
 import 'package:periodik/screens/point/point_dialog.dart';
 
 class SignalList extends ConsumerWidget {
@@ -25,6 +26,11 @@ class SignalList extends ConsumerWidget {
               point: point,
             );
           },
+          onLongPress: () => deletePoint(
+            context: context,
+            signalId: id,
+            pointId: point.id,
+          ),
           trailing: point.state
               ? Icon(
                   Icons.circle,